requirements.txt
spotipy
yt-dlp
ffmpeg-python
mutagen
fastapi
uvicorn
python-multipart
rich
python-dotenv